requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.16
albumentations==1.4.10
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.3.0
autopage==0.5.2
bitsandbytes==0.45.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cliff==4.8.0
cligj==0.7.2
cloudpickle==3.1.0
cmaes==0.11.1
cmd2==2.5.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.40.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
distlib==0.3.9
docker==7.1.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.0
executing==2.1.0
filelock==3.16.1
fiona==1.10.1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.27.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.3
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.35.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightly==1.5.15
lightly-utils==0.0.2
lightning==2.4.0
lightning-utilities==0.11.9
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.19.0
mlflow-skinny==2.19.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
optuna==2.10.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pretrainedmodels==0.7.4
prettytable==3.12.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pynvml==12.0.0
pyogrio==0.10.0
pyparsing==3.2.0
pyperclip==1.9.0
pyproj==3.7.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
requests==2.32.3
rich==13.9.4
rioxarray==0.18.1
rootutils==1.0.7
rsa==4.9
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
segmentation-models-pytorch==0.3.4
shapely==2.0.6
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terratorch==0.99.7
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.9.7
torch==2.5.0
torchgeo==0.6.2
torchmetrics==1.6.0
torchvision==0.20.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
xarray==2024.11.0
yarl==1.18.3
zipp==3.21.0
